refactor: components 폴더와 sections 폴더 분리
