+++ src/features/landingPage/sections/EarthSection.tsx
@@ -1,5 +1,5 @@
 import { useRef, useEffect, useState } from 'react';
-import EarthScene from '../EarthScene';
+import EarthScene from '../components/EarthScene';
 
 const EarthSection = () => {
   const earthSectionRef = useRef<HTMLDivElement>(null);
@@ -13,7 +13,7 @@ const EarthSection = () => {
   }, []);
 
   return (
-    <div ref={earthSectionRef} className="h-screen w-full bg-white overflow-hidden relative">
+    <div ref={earthSectionRef} className="h-[100dvh] w-full bg-white overflow-hidden relative">
       {triggerReady && <EarthScene earthAnimationTrigger={earthSectionRef} />}
     </div>
   );
